more cleanup for types and util
